Move some functions from util to cli_utils
