+++ k-distribution/src/main/scripts/lib/pyk/cli_utils.py
@@ -1,4 +1,10 @@
+import sys
+import time
+from datetime import datetime
 from pathlib import Path
+
+
+_lastTime = time.time()
 
 
 def check_dir_path(path: Path) -> None:
@@ -25,3 +31,25 @@ def file_path(s: str) -> Path:
     path = Path(s)
     check_file_path(path)
     return path
+
+
+def notif(msg):
+    global _lastTime
+    curTime   = time.time()
+    diffTime  = curTime - _lastTime
+    _lastTime = curTime
+    sys.stderr.write('== ' + sys.argv[0].split('/')[-1] + ' [+' + '{0:8.2f}'.format(diffTime) + ']: ' + msg + '\n')
+    sys.stderr.flush()
+
+
+def warning(msg):
+    notif('[WARNING] ' + msg)
+
+
+def fatal(msg, exitCode = 1):
+    notif('[FATAL] ' + msg)
+    raise('Quitting')
+
+
+def genFileTimestamp(comment = '//'):
+    return comment + ' This file generated by: ' + sys.argv[0] + '\n' + comment + ' ' + str(datetime.now()) + '\n'
