+++ k-distribution/src/main/scripts/lib/pyk/util.py
@@ -1,11 +1,7 @@
 #!/usr/bin/env python3
 
-from   datetime import datetime
 import hashlib
-import sys
-import time
 
-_lastTime = time.time()
 
 def combineDicts(*dicts):
     if len(dicts) == 0:
@@ -47,24 +43,6 @@ def dedupe(ls):
         if l not in newL:
             newL.append(l)
     return newL
-
-def notif(msg):
-    global _lastTime
-    curTime   = time.time()
-    diffTime  = curTime - _lastTime
-    _lastTime = curTime
-    sys.stderr.write('== ' + sys.argv[0].split('/')[-1] + ' [+' + '{0:8.2f}'.format(diffTime) + ']: ' + msg + '\n')
-    sys.stderr.flush()
-
-def warning(msg):
-    notif('[WARNING] ' + msg)
-
-def fatal(msg, exitCode = 1):
-    notif('[FATAL] ' + msg)
-    raise('Quitting')
-
-def genFileTimestamp(comment = '//'):
-    return comment + ' This file generated by: ' + sys.argv[0] + '\n' + comment + ' ' + str(datetime.now()) + '\n'
 
 def getAppliedAxiomList(debugLogFile):
     axioms      = []
